An attempt at fixing fill_paragraph inside comments
This adds invisible parens around consecutive line comments, so they'd be
treated as an s-exp by paredit-like commands (`forward_sexp`, `mark_sexp`,
`kill_sexp`, etc.). `fill_paragraph` will now check if we're inside a
comment, and if so, limit the fill region to it. It's not Emacs-perfect, but
better than before.